# saNotesWeb/src/main/resources/sanotes-openapi.yaml
openapi: 3.0.1
info:
  title: saNotes API
  description: Note Application
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/users/{username}:
    put:
      tags:
      - user
      summary: Update user
      operationId: updateUser
      parameters:
      - name: username
        in: path
        description: Username
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
      security:
      - bearerAuth: []
    delete:
      tags:
      - user
      summary: Delete user
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        description: Username
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearerAuth: []
  /api/users/{username}/removeAdmin:
    put:
      tags:
      - user
      summary: Remove admin role
      operationId: removeAdmin
      parameters:
      - name: username
        in: path
        description: Username
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearerAuth: []
  /api/users/{username}/giveAdmin:
    put:
      tags:
      - user
      summary: Update user as admin
      operationId: giveAdmin
      parameters:
      - name: username
        in: path
        description: Username
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearerAuth: []
...
  /api/users:
    post:
      tags:
      - user
      summary: Add user
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
      security:
      - bearerAuth: []
...
  /api/users/{username}/getUserItems:
    get:
      tags:
      - user
      summary: "Get user all items(notebooks, notes, tags)"
      operationId: getUserItems
      parameters:
      - name: username
        in: path
        description: Username
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - bearerAuth: []
  /api/users/isUsernameAvailable:
    get:
      tags:
      - user
      summary: Is username available
      description: Check for availability of username
      operationId: isUsernameAvailable
      parameters:
      - name: username
        in: query
        description: Username to check availability
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BooleanResponse'
  /api/users/isEmailAvailable:
    get:
      tags:
      - user
      summary: Is e-Mail available
      description: Check for availability of e-Mail
      operationId: isEmailAvailable
      parameters:
      - name: email
        in: query
        description: e-Mail to check availability
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExceptionResponse'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BooleanResponse'
...
components:
  schemas:
    ExceptionResponse:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
          format: int32
        messages:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    NoteBookModel:
      required:
      - description
      - name
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        notes:
          type: array
          items:
            $ref: '#/components/schemas/NotesModel'
        user:
          $ref: '#/components/schemas/User'
      description: Notebook parameters
    NotesModel:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        noteId:
          type: string
        topic:
          type: string
        text:
          type: string
        notebook:
          $ref: '#/components/schemas/NoteBookModel'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagModel'
      description: Note parameters
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          enum:
          - ROLE_ADMIN
          - ROLE_USER
    TagModel:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        notes:
          type: array
          items:
            $ref: '#/components/schemas/NotesModel'
        user:
          $ref: '#/components/schemas/User'
      description: Tag parameters
    User:
      required:
      - email
      - firstname
      - lastname
      - password
      - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstname:
          maxLength: 50
          minLength: 0
          type: string
        lastname:
          maxLength: 50
          minLength: 0
          type: string
        username:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 50
          minLength: 0
          type: string
        email:
          maxLength: 100
          minLength: 0
          type: string
          writeOnly: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        noteBooks:
          type: array
          items:
            $ref: '#/components/schemas/NoteBookModel'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagModel'
      description: User parameters
    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    NoteBookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        notes:
          type: array
          items:
            $ref: '#/components/schemas/NoteResponse'
    NoteResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        noteId:
          type: string
        topic:
          type: string
        text:
          type: string
        notebook:
          $ref: '#/components/schemas/NoteBookResponse'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstname:
          type: string
        lastname:
          type: string
        username:
          type: string
        email:
          type: string
        notebooks:
          type: array
          items:
            $ref: '#/components/schemas/NoteBookResponse'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
    BooleanResponse:
      type: object
      properties:
        result:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT